# docs/swagger.yaml
basePath: /
definitions:
...
  service.LoginForm:
    properties:
      isBak:
        type: boolean
      ledgerName:
        type: string
      openingBalances:
        type: string
      operatingCurrency:
        type: string
      secret:
        type: string
      startDate:
        type: string
    required:
    - ledgerName
    type: object
...
  service.UpdateConfigForm:
    properties:
      dataPath:
        type: string
      isBak:
        type: boolean
      openingBalances:
        type: string
      operatingCurrency:
        type: string
      secret:
        type: string
      startDate:
        type: string
    required:
    - dataPath
    - isBak
    - openingBalances
    - operatingCurrency
    - secret
    - startDate
    type: object
...
  service.UpdateSourceFileForm:
    properties:
      content:
        type: string
      path:
        type: string
    required:
    - path
    type: object
host: localhost:10000
info:
  contact:
    name: BaoXuebin
  description: Beancount-GS 是一个基于 Beancount 的个人财务管理系统，提供账本管理、交易记录、统计分析等功能。
  title: Beancount-GS API
  version: "1.0"
paths:
...
  /api/auth/file:
    post:
      consumes:
      - application/json
      description: 更新指定账本文件的内容，先自动备份再写入
      parameters:
      - description: 文件更新表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateSourceFileForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 更新账本文件内容
      tags:
      - 文件
  /api/auth/file/content:
    get:
      description: 返回指定相对路径的 .bean 源文件内容
      parameters:
      - description: 文件相对路径
        in: query
        name: path
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 读取账本文件内容
      tags:
      - 文件
  /api/auth/file/dir:
    get:
      description: 返回账本数据目录下所有 .bean 文件的相对路径列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取账本文件目录结构
      tags:
      - 文件
...
  /api/auth/ledger:
    delete:
      description: 删除当前账本的所有数据文件（不可恢复），需要 ledgerId 认证
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 删除账本
      tags:
      - 账本
  /api/auth/ledger/check:
    get:
      description: 运行 bean-check 对账本文件进行语法和余额校验，返回错误列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 检查账本合法性
      tags:
      - 账本
...
  /api/config:
    get:
      description: 返回服务器的全局配置信息，包含 DataPath、货币、日期等设置
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取全局服务配置
      tags:
      - 配置
    post:
      consumes:
      - application/json
      description: 更新服务器全局配置；需要提供 secret、startDate、dataPath、operatingCurrency 等字段
      parameters:
      - description: 全局配置表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateConfigForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 更新全局服务配置
      tags:
      - 配置
  /api/ledger:
    get:
      description: 返回所有已创建的账本信息列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取账本列表
      tags:
      - 账本
    post:
      consumes:
      - application/json
      description: 根据 ledgerName 和 secret 打开已有账本或创建新账本，成功后返回 ledgerId
      parameters:
      - description: 账本登录/创建表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.LoginForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 打开或创建账本
      tags:
      - 账本
...
securityDefinitions:
  LedgerId:
    description: 需要授权的接口请在 Header 中传入 ledgerId
    in: header
    name: ledgerId
    type: apiKey
swagger: "2.0"